api/requirements.txt
flask==2.2.2
#google-cloud-aiplatform==1.32.0
google-cloud-aiplatform[preview]
google-cloud-firestore==2.10.0
google-cloud-storage==2.10.0
Pillow==10.0.0
Flask-Cors==3.0.10
gunicorn==20.1.0
google.auth==2.16.0
requests==2.28.2
# opentelemetry libraries
opentelemetry-api==1.12.0
opentelemetry-sdk==1.12.0
opentelemetry-distro==0.33b0
opentelemetry-exporter-gcp-trace==1.3.0
opentelemetry-instrumentation==0.33b0
opentelemetry-instrumentation-flask==0.33b0
opentelemetry-resourcedetector-gcp==1.3.0a0
opentelemetry-semantic-conventions==0.33b0
opentelemetry-propagator-gcp==1.4.0
